requirements.txt
Flask
requests
scikit-learn
numpy
flask-limiter
python-dotenv
torch